first committ

+++ server/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModal');
+const User = require('../models/userModel');
+
 
 
 const getGoals = asyncHandler(async (req,res)=>{
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -18,7 +20,8 @@ const setGoal = asyncHandler(async (req,res)=>{
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal)
@@ -33,6 +36,20 @@ const updateGoal = asyncHandler(async (req,res)=>{
         res.status(400);
         throw new Error('Goal not found');
     }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal's user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorized for this operation')
+    }
+
     const updatedgoal = await Goal.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -49,6 +66,19 @@ const deleteGoal = asyncHandler(async (req,res)=>{
     if(!goal){
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id)
+
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal's user
+    if(goal.user.toString()!==user.id){
+        res.status(401)
+        throw new Error('User not authorized for this operation')
     }
 
     await goal.deleteOne()
